+++ CHANGELOG.md
@@ -5,6 +5,10 @@ all notable changes to this project will be documented in this file. the format 
 
 ## [unreleased]
 
+### changed
+
+- trim a lot of the about page's text
+
 ## [0.3.0] - 2025-07-04
 
 ### added