+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 all notable changes to this project will be documented in this file. the format is based on [keep a changelog](https://keepachangelog.com/en/1.1.0/) and adheres to [semantic versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.4.0] - 2025-07-12
 
 ### added